+++ src/slim-select/settings.ts
@@ -9,7 +9,6 @@ export default class Settings {
   public isMultiple: boolean = false
   public isOpen: boolean = false
   public isFullOpen: boolean = false
-  public intervalMove: ReturnType<typeof setInterval> | null = null
 
   // Fields set from constructor
   public disabled: boolean